+++ lib/pivotal_markdown.rb
@@ -1,6 +1,6 @@
-require "pivotal_markdown/version"
 require "pivotal_markdown/cli"
 require "pivotal_markdown/config"
-require "pivotal_markdown/error"
+require "pivotal_markdown/message"
+require "pivotal_markdown/version"
 
 require "pry"
